# api-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FutLendas API",
    "version": "1.0.0",
    "description": "Documentação da API backend para o aplicativo FutLendas."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Servidor de Desenvolvimento Local" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "CampeonatoUpdateDetailsInput": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "nome": { "type": "string" },
          "data": { "type": "string", "format": "date" }
        }
      },
...
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/campeonatos/{campeonato_id}": {
      "put": {
        "summary": "Atualizar detalhes do campeonato (Admin)",
        "description": "Atualiza o nome e/ou a data de um campeonato. Só é permitido se o campeonato ainda estiver na fase de 'inscricao'.",
        "tags": ["Campeonatos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "campeonato_id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CampeonatoUpdateDetailsInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Campeonato atualizado com sucesso." },
          "400": {
            "description": "Dados inválidos ou tentativa de editar um campeonato que não está mais na fase de 'inscricao'."
          },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado." },
          "404": { "description": "Campeonato não encontrado." }
        }
      },
      "delete": {
        "summary": "Deletar campeonato (Admin)",
        "description": "Remove permanentemente um campeonato e todos os seus dados associados (times inscritos, partidas, estatísticas) via ON DELETE CASCADE.",
        "tags": ["Campeonatos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "campeonato_id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Campeonato deletado com sucesso." },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado." },
          "404": { "description": "Campeonato não encontrado." }
        }
      }
    },
...
  },
...
}